requirements.txt
requests>=2.31
aiohttp>=3.9
rich>=13.0
tabulate>=0.9
python-dotenv>=1.0
